BUILDR-729 Integrate  into core and just make it the default pom generated.
