+++ CHANGELOG
@@ -7,6 +7,7 @@
 * Change: BUILDR-724 Use scalamain to call scala projects
 * Change: BUILDR-719 Change User-Agent when uploading artifacts
 * Change: Set the OPEN_IN_BROWSER to false when building idea launch targets via add_glassfish_remote_configuration.
+* Change: BUILDR-709 Integrate `buildr/custom_pom` into core and just make it the default pom generated.
 
 1.5.0 (2016-09-23)
 * Change: Update RJB to 1.5.4
