+++ spec/java/custom_pom_spec.rb
